+++ Frameworks/MoEngageSDK.xcframework/ios-arm64/MoEngageSDK.framework/CHANGELOG.md
@@ -3,6 +3,12 @@
 
 ## Changelog:
 
+### Version 9.16.0 *(27th February 2024)*
+-------------------------------------------
+* Refactored network layer.
+* Refactored data tracking.
+* Refactored logging.
+
 ### Version 9.15.2 *(9th February 2024)*
 -------------------------------------------
 * BugFix - Fixed network client leaking.